third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-4/styles-iframe.html

<html>
<head>

<style type="text/css" media="screen">
  #main { background:blue; }

/*# sourceURL=styles-iframe.html*/
</style>

<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/elements-test.js"></script>
<script>

function loadIframe()
{
    var iframe = document.createElement("iframe");
    iframe.src = "../styles/resources/styles-iframe-data.html";
    document.getElementById("main").appendChild(iframe);
}

function test() {
  TestRunner.evaluateInPage('loadIframe()');
  ConsoleTestRunner.addConsoleSniffer(step0);

  function step0() {
    ElementsTestRunner.selectNodeAndWaitForStyles('main', step1);
  }

  function step1() {
    TestRunner.addResult('Main frame style:');
    ElementsTestRunner.dumpSelectedElementStyles(true, false, true);
    ElementsTestRunner.selectNodeAndWaitForStyles('iframeBody', step2);
  }

  function step2() {
    TestRunner.addResult('iframe style:');
    ElementsTestRunner.dumpSelectedElementStyles(true, false, true);
    TestRunner.completeTest();
  }
}
</script>
</head>

<body onload="runTest()">
<p>
Tests that proper (and different) styles are returned for body elements of main document and iframe.
</p>

<div id="main"></div>
</body>
</html>
