third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-4/styles-iframe-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that proper (and different) styles are returned for body elements of main document and iframe.

Main frame style:
[expanded] 
element.style { ()

[expanded] 
#main { (styles-iframe.js:2 -> styles-iframe.js:2:16)
    background: blue;

[expanded] 
div { (user agent stylesheet)
    display: block;

iframe style:
[expanded] 
element.style { ()

[expanded] 
body { (styles-iframe-data.html:4 -> styles-iframe-data.html:4:9)
    background: green;

[expanded] 
body { (user agent stylesheet)
    display: block;
    margin: 8px;
